+++ package.json
@@ -21,6 +21,8 @@
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-emotion": "^8.0.12",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
@@ -54,6 +56,7 @@
     "react-router": "^4.2.0",
     "react-slick": "^0.16.0",
     "react-spinners": "^0.2.5",
-    "slick-carousel": "^1.8.1"
+    "slick-carousel": "^1.8.1",
+    "unistore": "^3.0.4"
   }
 }
