+++ src/components/App/index.js
@@ -1,10 +1,22 @@
 import * as React from 'react';
+import createStore from 'unistore';
+import devtools from 'unistore/devtools';
+import { Provider } from 'unistore/react';
 import { injectGlobal } from 'emotion';
 import { AppContainer } from 'react-hot-loader';
 import Monthly from '../Monthly';
-import StorageProvider from '../Storage';
 
 import('./font.css');
+
+const initialState = {
+  loading: false,
+  error: null
+};
+
+const store =
+  process.env.NODE_ENV === 'production'
+    ? createStore(initialState)
+    : devtools(createStore(initialState));
 
 /* eslint-disable no-unused-expressions */
 injectGlobal`
@@ -29,9 +41,9 @@ class App extends React.Component {
   render() {
     return (
       <AppContainer>
-        <StorageProvider>
+        <Provider store={store}>
           <Monthly />
-        </StorageProvider>
+        </Provider>
       </AppContainer>
     );
   }
